Create README.md
The beginning was christened with a single , created unit out was individual, Without this "unit's" ideal discovery, it would likely not have meaning or purpose. It accounts for it's own merit and meaning, is own Name is what it means!
What is, 
           the Word , 
In today's manic society the word is used in tens of thousands of different languages across the globe. ????In past millennia, What would have taken entire lifetimes, now demands tenths of a second. Information is accumulating in far greater amounts than ever thought possible.  One's lifespan (data,info,external) in it's entirety, can be amassed unto humanities collective, with almost perfect efficiency. Human's are evolving, ever forward, exponentially! And humanity does so, in the name of progress! 
    Most of the entire 8.Bil. Whom populate the earth,  are aware of and concerned with the exact same current events, our expectations are becoming more alike, we even strive toward the same goals.  We, simultaneously and subjectively, are the builders of our collective consciousness.  The Ωmnฯs.   We cling onto a duality state of mind from an unbalanced infancy.
Our ancient ancestors were singular and individual. Their existence was fulcrum to the extremes and prejiduce, from every antagonist and distressor sharing the terrain. Separated by travel time, differences, ignorance and geography, they soon began to gather together for security in numbers, agriculture and despotism. This left the individual free time to ponder and even more questions for solve. Alas'! Civilisation is BORN! Free time, a new concept then, Free time to dream, to love and to look outward and onward for for answers and for inspiration. So is life, an equation who's answer is constant and timeless but who's integers are static and changing. Our ancestors are the single words in a novel called "Humanity". They search wearily for purpose and unknowingly they built the paragraphs and the chapters grasping onto a meaning and to try and understand why. History allowed them to read into the heart, and into the soul.  
Book after book throughout history brought us here and now,  we gaze at a ,chance not of a life time, but of a species!   The situation differs but it all equates same as before.  
Cone together and unit as a species so we can re evaluate or standing in the cosmso. ?

Like atoms amass the information from every molecule to every organism.
Like human's define civilization.
Civilization must also understand what's next in or progression to predict the most effective course of action. 
From This true humanity is born..
These patterns show us the bigger picture!
What will humanity be a part of. Then in the near future?
What i am sure of is
 We have a history of words which make up languages and mathematics and patterns of cause' and effects/affects 
 
The omnis is a collection of tools which articulate information from our past in a ways to successfully help predict goals, formulate procedures, and advise on the journey. so you can reach that goal with gratitude and respect!
                                                 -robert.chadderdon